+++ include/mem/vmm.h
@@ -45,6 +45,6 @@ void vmm_unmap_region(uintptr_t virt_base, uint64_t size);
 void vmm_unmap_page(uintptr_t virt);
 uintptr_t vmm_get_phys(uintptr_t virt);
 void *vmm_map_phys(uint64_t addr, uint64_t len);
-void vmm_unmap_virt(void *addr, size_t len);
+void vmm_unmap_virt(void *addr, uint64_t len);
 extern struct page_table *kernel_pml4;
 #pragma once
