+++ kernel/mem/pmalloc.c
@@ -11,15 +11,15 @@
 
 static uint8_t bitmap[BITMAP_SIZE];
 
-static void set_bit(size_t index) {
+static void set_bit(uint64_t index) {
     bitmap[index / 8] |= (1 << (index % 8));
 }
 
-static void clear_bit(size_t index) {
+static void clear_bit(uint64_t index) {
     bitmap[index / 8] &= ~(1 << (index % 8));
 }
 
-static bool test_bit(size_t index) {
+static bool test_bit(uint64_t index) {
     return (bitmap[index / 8] & (1 << (index % 8))) != 0;
 }
 
@@ -45,7 +45,7 @@ void init_physical_allocator(uint64_t o, struct limine_memmap_request m) {
             uint64_t end = (entry->base + entry->length) & ~(PAGE_SIZE - 1);
 
             for (uint64_t addr = start; addr < end; addr += PAGE_SIZE) {
-                size_t index = addr / PAGE_SIZE;
+                uint64_t index = addr / PAGE_SIZE;
                 if (index < BITMAP_SIZE * 8) {
                     clear_bit(index);
                 }
@@ -56,7 +56,7 @@ void init_physical_allocator(uint64_t o, struct limine_memmap_request m) {
 
 void *pmm_alloc_page(bool add_offset) {
 
-    for (size_t i = 0; i < BITMAP_SIZE * 8; i++) {
+    for (uint64_t i = 0; i < BITMAP_SIZE * 8; i++) {
         if (!test_bit(i)) {
             set_bit(i);
             void *page = (void *) ((add_offset ? offset : 0) + (i * PAGE_SIZE));
@@ -72,11 +72,11 @@ void *pmm_alloc_page(bool add_offset) {
  * Used for debug and log.
  */
 void print_memory_status() {
-    size_t total_pages = BITMAP_SIZE * 8;
-    size_t free_pages = 0;
-    size_t allocated_pages = 0;
+    uint64_t total_pages = BITMAP_SIZE * 8;
+    uint64_t free_pages = 0;
+    uint64_t allocated_pages = 0;
 
-    for (size_t i = 0; i < total_pages; i++) {
+    for (uint64_t i = 0; i < total_pages; i++) {
         if (!test_bit(i)) {
             free_pages++;
         } else {
@@ -92,10 +92,10 @@ void print_memory_status() {
 
     k_printf("\nMemory Segments (contiguous):\n");
 
-    size_t segment_start = 0;
+    uint64_t segment_start = 0;
     int segment_state = test_bit(0);
 
-    for (size_t i = 1; i <= total_pages; i++) {
+    for (uint64_t i = 1; i <= total_pages; i++) {
         int current_state = (i < total_pages) ? test_bit(i) : -1;
         if (current_state != segment_state) {
 
@@ -117,7 +117,7 @@ void print_memory_status() {
 /*
  * Allocate `count` pages.
  */
-void *pmm_alloc_pages(size_t count, bool add_offset) {
+void *pmm_alloc_pages(uint64_t count, bool add_offset) {
 
     if (count == 0) {
         return NULL;
@@ -127,11 +127,11 @@ void *pmm_alloc_pages(size_t count, bool add_offset) {
         return pmm_alloc_page(add_offset);
     }
 
-    size_t consecutive = 0;
-    size_t start_index = 0;
+    uint64_t consecutive = 0;
+    uint64_t start_index = 0;
     bool found = false;
 
-    for (size_t i = 0; i < BITMAP_SIZE * 8; i++) {
+    for (uint64_t i = 0; i < BITMAP_SIZE * 8; i++) {
 
         if (!test_bit(i)) {
             if (consecutive == 0) {
@@ -153,7 +153,7 @@ void *pmm_alloc_pages(size_t count, bool add_offset) {
         return NULL;
     }
 
-    for (size_t i = 0; i < count; i++) {
+    for (uint64_t i = 0; i < count; i++) {
         set_bit(start_index + i);
     }
 
@@ -165,24 +165,24 @@ void *pmm_alloc_pages(size_t count, bool add_offset) {
  *
  * Addresses should have the HHDM offset added to them.
  */
-void pmm_free_pages(void *addr, size_t count, bool has_offset) {
+void pmm_free_pages(void *addr, uint64_t count, bool has_offset) {
 
     if (addr == NULL || count == 0) {
         return;
     }
 
-    size_t start_index =
-        ((size_t) addr - (has_offset ? offset : 0)) / PAGE_SIZE;
+    uint64_t start_index =
+        ((uint64_t) addr - (has_offset ? offset : 0)) / PAGE_SIZE;
 
     if (start_index >= BITMAP_SIZE * 8 ||
         start_index + count > BITMAP_SIZE * 8) {
         k_printf("Invalid address range to free: 0x%zx with count %zu\n",
-                 (size_t) addr, count);
+                 (uint64_t) addr, count);
         return;
     }
 
-    for (size_t i = 0; i < count; i++) {
-        size_t index = start_index + i;
+    for (uint64_t i = 0; i < count; i++) {
+        uint64_t index = start_index + i;
         if (test_bit(index)) {
             clear_bit(index);
         } else {
