+++ kernel/mem/vmm.c
@@ -269,8 +269,8 @@ void *vmm_map_phys(uint64_t addr, uint64_t len) {
     uintptr_t phys_start = PAGE_ALIGN_DOWN(addr);
     uintptr_t offset = addr - phys_start;
 
-    size_t total_len = len + offset;
-    size_t total_pages = (total_len + PAGE_SIZE - 1) / PAGE_SIZE;
+    uint64_t total_len = len + offset;
+    uint64_t total_pages = (total_len + PAGE_SIZE - 1) / PAGE_SIZE;
 
     if (vmm_map_top + total_pages * PAGE_SIZE > VMM_MAP_LIMIT) {
         return NULL;
@@ -279,7 +279,7 @@ void *vmm_map_phys(uint64_t addr, uint64_t len) {
     uintptr_t virt_start = vmm_map_top;
     vmm_map_top += total_pages * PAGE_SIZE;
 
-    for (size_t i = 0; i < total_pages; i++) {
+    for (uint64_t i = 0; i < total_pages; i++) {
         vmm_map_page(virt_start + i * PAGE_SIZE, phys_start + i * PAGE_SIZE,
                      PAGING_PRESENT | PAGING_WRITE);
     }
@@ -287,15 +287,15 @@ void *vmm_map_phys(uint64_t addr, uint64_t len) {
     return (void *) (virt_start + offset);
 }
 
-void vmm_unmap_virt(void *addr, size_t len) {
+void vmm_unmap_virt(void *addr, uint64_t len) {
     uintptr_t virt_addr = (uintptr_t) addr;
     uintptr_t page_offset = virt_addr & (PAGE_SIZE - 1);
     uintptr_t aligned_virt = PAGE_ALIGN_DOWN(virt_addr);
 
-    size_t total_len = len + page_offset;
-    size_t total_pages = (total_len + PAGE_SIZE - 1) / PAGE_SIZE;
+    uint64_t total_len = len + page_offset;
+    uint64_t total_pages = (total_len + PAGE_SIZE - 1) / PAGE_SIZE;
 
-    for (size_t i = 0; i < total_pages; i++) {
+    for (uint64_t i = 0; i < total_pages; i++) {
         vmm_unmap_page(aligned_virt + i * PAGE_SIZE);
     }
 }
